refactor: file splitting & code cleanup
