+++ src/components/backlog-table/backlog-table-row-actions.tsx
@@ -1,0 +1,87 @@
+import type { Row } from '@tanstack/react-table';
+import { MoreHorizontal } from 'lucide-react';
+
+import type { BacklogItem } from '@/apollo/types/types';
+import { Button } from '@/ui/button';
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from '@/ui/dialog-alert';
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from '@/ui/dropdown-menu';
+import { Sheet } from '@/ui/sheet';
+
+type BacklogTableRowActionsProps = {
+  row: Row<BacklogItem>;
+  onFormEditClick: (row: Row<BacklogItem>) => VoidFunction;
+  onRowDeleteClick: (row: Row<BacklogItem>) => VoidFunction;
+};
+
+export function BacklogTableRowActions({
+  row,
+  onFormEditClick,
+  onRowDeleteClick,
+}: BacklogTableRowActionsProps) {
+  const { task } = row.original;
+
+  return (
+    <Sheet>
+      <DropdownMenu>
+        <DropdownMenuTrigger asChild>
+          <Button variant="ghost" className="h-8 w-8 p-0 cursor-pointer">
+            <span className="sr-only">Open menu</span>
+            <MoreHorizontal />
+          </Button>
+        </DropdownMenuTrigger>
+        <DropdownMenuContent align="end">
+          <DropdownMenuLabel>Actions</DropdownMenuLabel>
+          <DropdownMenuSeparator />
+          <DropdownMenuItem
+            className="cursor-pointer"
+            onClick={onFormEditClick(row)}
+          >
+            Edit
+          </DropdownMenuItem>
+          <AlertDialog>
+            <AlertDialogTrigger asChild>
+              <DropdownMenuItem
+                className="text-red-600 cursor-pointer hover:text-red-600!"
+                onSelect={(e) => e.preventDefault()}
+              >
+                Delete
+              </DropdownMenuItem>
+            </AlertDialogTrigger>
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                <AlertDialogTitle>Delete task {`"${task}"`}?</AlertDialogTitle>
+                <AlertDialogDescription>
+                  This action cannot be undone. This will permanently delete
+                  backlog item.
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+              <AlertDialogFooter>
+                <AlertDialogCancel>Cancel</AlertDialogCancel>
+                <AlertDialogAction onClick={onRowDeleteClick(row)}>
+                  Continue
+                </AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
+        </DropdownMenuContent>
+      </DropdownMenu>
+    </Sheet>
+  );
+}
